# docs/swagger.yaml
basePath: /api/v0
definitions:
  types.Collection:
    properties:
      backdrop_path:
        type: string
      id:
        type: integer
      name:
        type: string
      poster_path:
        type: string
    type: object
  types.Company:
    properties:
      id:
        type: integer
      logo_path:
        type: string
      name:
        type: string
      origin_country:
        type: string
    type: object
  types.Country:
    properties:
      iso_3166_1:
        type: string
      name:
        type: string
    type: object
  types.Creator:
    properties:
      credit_id:
        type: string
      gender:
        type: integer
      id:
        type: integer
      name:
        type: string
      profile_path:
        type: string
    type: object
  types.Episode:
    properties:
      air_date:
        type: string
      episode_number:
        type: integer
      id:
        type: integer
      name:
        type: string
      overview:
        type: string
      production_code:
        type: string
      runtime:
        type: integer
      season_number:
        type: integer
      still_path:
        type: string
      vote_average:
        type: number
      vote_count:
        type: integer
    type: object
...
  types.ExternalIds:
    properties:
      facebook_id:
        type: string
      freebase_id:
        type: string
      freebase_mid:
        type: string
      imdb_id:
        type: string
      instagram_id:
        type: string
      tvdb_id:
        type: integer
      tvrage_id:
        type: integer
      twitter_id:
        type: string
    type: object
  types.Genre:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  types.Language:
    properties:
      iso_639_1:
        type: string
      name:
        type: string
    type: object
...
  types.Movie:
    properties:
      adult:
        type: boolean
      backdrop_path:
        type: string
      genre_ids:
        items:
          type: integer
        type: array
      id:
        type: integer
      original_language:
        type: string
      original_title:
        type: string
      overview:
        type: string
      popularity:
        type: number
      poster_path:
        type: string
      release_date:
        type: string
      title:
        type: string
      video:
        type: boolean
      vote_average:
        type: number
      vote_count:
        type: integer
    type: object
  types.MovieInfo:
    properties:
      adult:
        type: boolean
      backdrop_path:
        type: string
      belongs_to_collection:
        items:
          $ref: '#/definitions/types.Collection'
        type: array
      budget:
        type: integer
      genres:
        items:
          $ref: '#/definitions/types.Genre'
        type: array
      homepage:
        type: string
      id:
        type: integer
      imdb_id:
        type: string
      original_language:
        type: string
      original_title:
        type: string
      overview:
        type: string
      popularity:
        type: number
      poster_path:
        type: string
      production_companies:
        items:
          $ref: '#/definitions/types.Company'
        type: array
      production_countries:
        items:
          $ref: '#/definitions/types.Country'
        type: array
      release_date:
        type: string
      revenue:
        type: integer
      runtime:
        type: integer
      spoken_languages:
        items:
          $ref: '#/definitions/types.Language'
        type: array
      status:
        type: string
      tagline:
        type: string
      title:
        type: string
      video:
        type: boolean
      vote_average:
        type: number
      vote_count:
        type: integer
    type: object
  types.MovieResults:
    properties:
      page:
        type: integer
      results:
        items:
          $ref: '#/definitions/types.Movie'
        type: array
      total_pages:
        type: integer
      total_results:
        type: integer
    type: object
...
  types.Season:
    properties:
      air_date:
        type: string
      episode_count:
        type: integer
      id:
        type: integer
      name:
        type: string
      overview:
        type: string
      poster_path:
        type: string
      season_number:
        type: integer
    type: object
  types.Show:
    properties:
      backdrop_path:
        type: string
      first_air_date:
        type: string
      genre_ids:
        items:
          type: integer
        type: array
      id:
        type: integer
      name:
        type: string
      origin_country:
        items:
          type: string
        type: array
      original_language:
        type: string
      original_name:
        type: string
      overview:
        type: string
      popularity:
        type: number
      poster_path:
        type: string
      vote_average:
        type: number
      vote_count:
        type: integer
    type: object
  types.ShowInfo:
    properties:
      adult:
        type: boolean
      backdrop_path:
        type: string
      created_by:
        items:
          $ref: '#/definitions/types.Creator'
        type: array
      episode_run_time:
        items:
          type: integer
        type: array
      external_ids:
        $ref: '#/definitions/types.ExternalIds'
      first_air_date:
        type: string
      genres:
        items:
          $ref: '#/definitions/types.Genre'
        type: array
      homepage:
        type: string
      id:
        type: integer
      in_production:
        type: boolean
      languages:
        items:
          type: string
        type: array
      last_air_date:
        type: string
      last_episode_to_air:
        $ref: '#/definitions/types.Episode'
      name:
        type: string
      networks:
        items:
          $ref: '#/definitions/types.Company'
        type: array
      next_episode_to_air:
        $ref: '#/definitions/types.Episode'
      number_of_episodes:
        type: integer
      number_of_seasons:
        type: integer
      origin_country:
        items:
          type: string
        type: array
      original_language:
        type: string
      original_name:
        type: string
      overview:
        type: string
      popularity:
        type: number
      poster_path:
        type: string
      production_companies:
        items:
          $ref: '#/definitions/types.Company'
        type: array
      production_countries:
        items:
          $ref: '#/definitions/types.Country'
        type: array
      seasons:
        items:
          $ref: '#/definitions/types.Season'
        type: array
      spoken_languages:
        items:
          $ref: '#/definitions/types.Language'
        type: array
      status:
        type: string
      tagline:
        type: string
      type:
        type: string
      vote_average:
        type: number
      vote_count:
        type: integer
    type: object
  types.ShowResults:
    properties:
      page:
        type: integer
      results:
        items:
          $ref: '#/definitions/types.Show'
        type: array
      total_pages:
        type: integer
      total_results:
        type: integer
    type: object
...
  types.TorrentFile:
    properties:
      length:
        type: string
      name:
        type: string
      url:
        type: string
    type: object
...
  v0.MessageResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  v0.TmdbMovieInfoResponse:
    properties:
      results:
        $ref: '#/definitions/types.MovieInfo'
      success:
        type: boolean
    type: object
  v0.TmdbMovieResultsResponse:
    properties:
      results:
        $ref: '#/definitions/types.MovieResults'
      success:
        type: boolean
    type: object
  v0.TmdbShowInfoResponse:
    properties:
      results:
        $ref: '#/definitions/types.ShowInfo'
      success:
        type: boolean
    type: object
  v0.TmdbShowResultsResponse:
    properties:
      results:
        $ref: '#/definitions/types.ShowResults'
      success:
        type: boolean
    type: object
  v0.TorrentFilesResultsResponse:
    properties:
      hash:
        type: string
      results:
        items:
          $ref: '#/definitions/types.TorrentFile'
        type: array
      success:
        type: boolean
    type: object
...
host: localhost:9000
info:
  contact: {}
  title: Raven Torrent API
  version: 0.6.0
paths:
...
  /add/{base64uri}:
    get:
      parameters:
      - description: Link to torrent file / magnet link (base64 encoded)
        in: path
        name: base64uri
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.TorrentFilesResultsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Get torrent info and streaming URLs
      tags:
      - Torrent client
  /cast/{base64location}/{base64query}:
    get:
      parameters:
      - description: ' '
        in: path
        name: base64location
        required: true
        type: string
      - description: ' '
        in: path
        name: base64query
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Cast media file to TV or media player
      tags:
      - Media playback
  /delete/{hash}:
    get:
      parameters:
      - description: ' '
        in: path
        name: hash
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Delete torrent from torrent client
      tags:
      - Torrent client
  /deleteall:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Delete all torrents from torrent client
      tags:
      - Torrent client
...
  /tmdbdiscover/type/movie/genretype/{genretype}/sort/{sort}/date/{date}/lang/{lang}/page/{page}:
    get:
      parameters:
      - description: ' '
        in: path
        name: genretype
        required: true
        type: string
      - description: ' '
        in: path
        name: sort
        required: true
        type: string
      - description: ' '
        in: path
        name: date
        required: true
        type: string
      - description: ' '
        in: path
        name: lang
        required: true
        type: string
      - description: ' '
        in: path
        name: page
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.TmdbMovieResultsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Discover movies by genre
      tags:
      - Media search
  /tmdbdiscover/type/tv/genretype/{genretype}/sort/{sort}/date/{date}/lang/{lang}/page/{page}:
    get:
      parameters:
      - description: ' '
        in: path
        name: genretype
        required: true
        type: string
      - description: ' '
        in: path
        name: sort
        required: true
        type: string
      - description: ' '
        in: path
        name: date
        required: true
        type: string
      - description: ' '
        in: path
        name: lang
        required: true
        type: string
      - description: ' '
        in: path
        name: page
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.TmdbShowResultsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Discover shows by genre
      tags:
      - Media search
  /tmdbinfo/type/movie/tmdbid/{tmdbid}/lang/{lang}:
    get:
      parameters:
      - description: ' '
        in: path
        name: tmdbid
        required: true
        type: string
      - description: ' '
        in: path
        name: lang
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.TmdbMovieInfoResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Get movie details
      tags:
      - Media search
  /tmdbinfo/type/tv/tmdbid/{tmdbid}/lang/{lang}:
    get:
      parameters:
      - description: ' '
        in: path
        name: tmdbid
        required: true
        type: string
      - description: ' '
        in: path
        name: lang
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.TmdbShowInfoResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Get show details
      tags:
      - Media search
...
swagger: "2.0"